+++ example/oglplus/029_fire_hydrant.cpp
@@ -211,8 +211,8 @@ public:
 			float u = std::rand() / float(RAND_MAX);
 			float v = std::rand() / float(RAND_MAX);
 			shadow_offs[i].Set(
-				std::sqrt(v) * std::cos(2*3.1415*u),
-				std::sqrt(v) * std::sin(2*3.1415*u),
+				float(std::sqrt(v) * std::cos(2*3.1415*u)),
+				float(std::sqrt(v) * std::sin(2*3.1415*u)),
 				0.0f
 			);
 		}
